requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.196
pydantic==2.5.0
python-multipart==0.0.6